+++ src/utils/codecs/bin/mod.rs
@@ -6,11 +6,11 @@ mod msgpack_serde;
 #[cfg(feature = "prost")]
 mod prost;
 
-#[cfg(feature = "bincode")]
+#[cfg(feature = "bincode_serde")]
 pub use bincode_serde::*;
 #[allow(unused_imports)]
 pub use from_to_bytes::*;
-#[cfg(feature = "msgpack")]
+#[cfg(feature = "msgpack_serde")]
 pub use msgpack_serde::*;
 #[cfg(feature = "prost")]
 pub use prost::*;